+++ src/main/java/cs211/project/services/EventDataSource.java
@@ -44,8 +44,8 @@ public class EventDataSource implements DataSource<EventList>{
             while ((line = buffer.readLine()) != null) {
                 String[] data = line.split(",");
                 Event newEvent = new Event(data[0].trim(), Long.parseLong(data[1]), Long.parseLong(data[2]), data[3].trim(),
-                        Double.parseDouble(data[4]), Double.parseDouble(data[5]), Double.parseDouble(data[6]), data[7].trim(),
-                        Double.parseDouble(data[8]), data[9].trim(), data[10].trim(), data[11]);
+                        Long.parseLong(data[4]), Long.parseLong(data[5]), Long.parseLong(data[6]), data[7].trim(),
+                        Long.parseLong(data[8]), data[9].trim(), data[10].trim(), data[11]);
                 eventList.addEvent(newEvent);
             }
         } catch (FileNotFoundException e) {
